pyproject.toml
[project]
name = "smart-student-assistent"
version = "0.1.0"
description = "Smart Student Assistant"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
]
